requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
scipy>=1.10
liac-arff>=2.5.0
matplotlib>=3.7
joblib>=1.3


